zdtm: add support for the "freezecg" option in the .desc file
Some tests may need the freezecg option to run.
The only way to do this is to pass this option to zdtm.py.
So these tests can't be run automatically.

'freezecg' declaration in the .desc file allows running these
tests with the "-a" command.

Let's add support for the "freezecg" in the .desc file.
As well as the zdtm.py option, the .desc option shouldn't
be run in parallel.

Signed-off-by: Yuriy Vasiliev <yuriy.vasiliev@virtuozzo.com>